+++ packages/yourloops/test/integration/mock/team.api.mock.ts
@@ -48,12 +48,15 @@ export const monitoringParameters = {
   reportingPeriod: 7
 }
 
+export const myTeamId = '62fe00defc9374b2fed32bcc'
 export const mySecondTeamId = '62fe00defc9374b2fed32bcd'
 export const myThirdTeamId = '62fe00defc9374b2fed32bce'
+export const mySecondTeamName = 'MySecondTeam'
+export const myThirdTeamName = 'MyThirdTeam - to be deleted'
 
 export const teamOne: Team = {
   name: 'MyTeam',
-  id: '62fe00defc9374b2fed32bcc',
+  id: myTeamId,
   code: '036038775',
   type: TeamType.medical,
   phone: '0476000000',
@@ -76,7 +79,7 @@ export const teamOne: Team = {
 }
 
 export const teamTwo: Team = {
-  name: 'MySecondTeam',
+  name: mySecondTeamName,
   id: mySecondTeamId,
   code: '241548778',
   type: TeamType.medical,
@@ -106,7 +109,7 @@ export const teamTwo: Team = {
 }
 
 export const teamThree: Team = {
-  name: 'MyThirdTeam - to be deleted',
+  name: myThirdTeamName,
   id: myThirdTeamId,
   code: '263381988',
   type: TeamType.medical,