+++ packages/yourloops/test/integration/pages/patient-dashboard/patient-dashboard-common.spec.tsx
@@ -27,7 +27,7 @@
 
 import { mockAuth0Hook } from '../../mock/auth0.hook.mock'
 import { mockTeamAPI } from '../../mock/team.api.mock'
-import { mockDataAPI } from '../../mock/data.api.mock'
+import { completeDashboardData, mockDataAPI } from '../../mock/data.api.mock'
 import { mockNotificationAPI } from '../../mock/notification.api.mock'
 import { mockPatientAPI, unmonitoredPatientId } from '../../mock/patient.api.mock'
 import { mockChatAPI } from '../../mock/chat.api.mock'
@@ -37,6 +37,15 @@ import { renderPage } from '../../utils/render'
 import { screen, within } from '@testing-library/react'
 import { checkNoTooltip } from '../../assert/stats'
 import { mockUserApi } from '../../mock/user.api.mock'
+import crypto from 'crypto'
+
+// window.crypto is not defined in jest...
+Object.defineProperty(global, 'crypto', {
+  value: {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    getRandomValues: (arr: any) => crypto.randomBytes(arr.length)
+  }
+})
 
 describe('Patient dashboard for anyone', () => {
   const firstName = 'HCP firstName'
@@ -51,10 +60,10 @@ describe('Patient dashboard for anyone', () => {
     mockPatientAPI()
     mockChatAPI()
     mockMedicalFilesAPI()
-    mockDataAPI()
+    mockDataAPI(completeDashboardData)
   })
 
-  it('should not render tooltips on patient statistics', async () => {
+  it('statistics should be displayed without tooltips and with correct labels', async () => {
     renderPage(`/patient/${unmonitoredPatientId}/dashboard`)
     const statsWidgets = within(await screen.findByTestId('patient-statistics', {}, { timeout: 3000 }))
     checkNoTooltip(statsWidgets, 'Time In Range')
@@ -62,5 +71,15 @@ describe('Patient dashboard for anyone', () => {
     checkNoTooltip(statsWidgets, 'Avg. Daily Insulin')
     checkNoTooltip(statsWidgets, 'Avg. Daily Time In Loop Mode')
     checkNoTooltip(statsWidgets, 'Avg. Daily Carbs')
+
+    const deviceUsageWidget = within(screen.getByTestId('device-usage-card'))
+    checkNoTooltip(deviceUsageWidget, 'Sensor Usage')
+
+    expect(statsWidgets.getByTestId('stat-timeInRange')).toHaveTextContent('Time In Range2h8%10h42%6h25%4h17%2h8%<5454-7070-180180-250>250mg/dL')
+    expect(statsWidgets.getByTestId('cbg-mean-stat')).toHaveTextContent('Avg. Glucose (CGM)mg/dL135')
+    expect(statsWidgets.getByTestId('stat-averageDailyDose')).toHaveTextContent('Avg. Daily Insulin1.3UWeight72kgDaily Dose ÷ Weight0.02U/kg')
+    expect(statsWidgets.getByTestId('loop-mode-stat')).toHaveTextContent('Avg. Daily Time In Loop ModeONOFF91%21h 49m9%2h 11m')
+    expect(statsWidgets.getByTestId('total-carbs-stat')).toHaveTextContent('Avg. Daily Carbs55gRescue carbs10g')
+    expect(deviceUsageWidget.getByTestId('stat-sensorUsage')).toHaveTextContent('Sensor Usage1%')
   })
 })