+++ packages/yourloops/test/unit/components/patient/patient-row.test.tsx
@@ -85,7 +85,7 @@ describe('Patient row', () => {
 
   it('should open modal when clicking on remove patient icon', () => {
     render(getPatientRowJSX())
-    const removeButton = screen.getByRole('button', { name: 'remove-patient-fake@email.com' })
+    const removeButton = screen.getByRole('button', { name: 'remove-patient fake@email.com' })
     expect(screen.queryByTestId(removePatientDialogMockId)).not.toBeInTheDocument()
     fireEvent.click(removeButton)
     expect(screen.queryByTestId(removePatientDialogMockId)).toBeInTheDocument()
@@ -120,7 +120,7 @@ describe('Patient row', () => {
     render(getPatientRowJSX())
     const cells = screen.getAllByRole('cell')
     expect(cells).toHaveLength(10)
-    expect(within(cells[1]).queryByText(patient.profile.fullName)).not.toBeNull()
+    expect(within(cells[1]).queryByText('user-name')).not.toBeNull()
     expect(within(cells[2]).queryByText('N/A')).not.toBeNull()
     expect(within(cells[3]).queryByText('no')).not.toBeNull()
     expect(within(cells[4]).queryByText(`${patient.metadata.alarm.timeSpentAwayFromTargetRate}%`)).not.toBeNull()
@@ -144,7 +144,7 @@ describe('Patient row', () => {
     render(getPatientRowJSX())
     const cells = screen.getAllByRole('cell')
     expect(cells).toHaveLength(8)
-    expect(within(cells[1]).queryByText(patient.profile.fullName)).not.toBeNull()
+    expect(within(cells[1]).queryByText('user-name')).not.toBeNull()
     expect(within(cells[2]).queryByText('N/A')).not.toBeNull()
     expect(within(cells[3]).queryByText(`${patient.metadata.alarm.timeSpentAwayFromTargetRate}%`)).not.toBeNull()
     expect(within(cells[4]).queryByText(`${patient.metadata.alarm.frequencyOfSevereHypoglycemiaRate}%`)).not.toBeNull()