+++ CHANGELOG.md
@@ -9,9 +9,10 @@ It is based on Tidepool Blip 1.27.
 - YLP-1850 Patient search should comply to INS standard
 - YLP-1943 Training material is acknowledged after Yourloops TM is opened and read (including the new checkbox message) and before any use of YLP
 - YLP-1944 Rename "Intended Use" in the footer as "Product labelling"
+- YLP-1947 Add team selection dropdown on patient dashboard for HCP
 - YLP-1973 User should be able to uncheck the training check box
 ### Engineering use
-- YLP-1976 Fix PDF generation
+- YLP-1444 Handle user name display for Japanese
 - YLP-1624 Fix error messages when inviting a patient already in the team
 - YLP-1625 Display an explicit error when trying to share data with another patient
 - YLP-1680 getPatientData Api method has a wrong parameter
@@ -34,8 +35,11 @@ It is based on Tidepool Blip 1.27.
 - YLP-1949 Migrate Lines component to dumb
 - YLP-1953 Migrate WheelPercent to dumb
 - YLP-1960 enhance pipeline
+- YLP-1968 Removing eCPS connection feature (revert of YLP-1182)
+- YLP-1976 Fix PDF generation
 - Automate copyright generation on new and updated files
 - Changing CI from Jenkins to GitHub Actions
+- Migrate MUI 4 to MUI 5
 - Refactoring profile page code and adding tests
 - Remove react-test-renderer dependency
 - Update copyright on all files in `dumb` and `yourloops` packages